+++ packages/ripple-tide-landing-page/src/mapping/page-components/webforms/webforms-mapping.ts
@@ -1,7 +1,16 @@
 import { TideDynamicPageComponent } from '@dpc-sdp/ripple-tide-api'
 import { FormKitSchemaNode } from '@formkit/core'
+import {
+  getCounterFields,
+  getInputIcons,
+  getMinMaxFields,
+  getValidation
+} from './webform-utils.js'
 
 export interface ITideWebform {
+  formId: string
+  successMessageHTML: string
+  errorMessageHTML: string
   schema: FormKitSchemaNode
 }
 
@@ -15,75 +24,10 @@ export interface ITideFormElement {
 }
 
 const getFormSchemaFromMapping = async (
-  elements: ITideFormElement[],
+  webform,
   tidePageApi
 ): Promise<FormKitSchemaNode[]> => {
-  const getValidation = (
-    field
-  ): {
-    validation: string
-    validationMessages: Record<string, string>
-    validationMeta: string | undefined
-  } => {
-    const validationStates: string[] = []
-    const validationMessages: Record<string, string> = {}
-    const validationMeta = undefined
-
-    // Validation states from Drupal
-    // Currently supports required, pattern
-    if (field['#required']) {
-      const requiredMessage =
-        field['#required_error'] || `${field['#title']} is required`
-
-      validationStates.push('required')
-      validationMessages['required'] = requiredMessage
-
-      // The formkit required rule accepts `false` as a value, so we need to use add the 'accepted' rule as well
-      if (
-        field['#type'] === 'checkbox' ||
-        field['#type'] === 'webform_privacy_statement'
-      ) {
-        validationStates.push('accepted')
-        validationMessages['accepted'] = requiredMessage
-      }
-    }
-    if (field['#pattern']) {
-      validationStates.push(`matches:${field['#pattern']}`)
-      validationMessages['matches'] =
-        field['#pattern_error'] ||
-        `${field['#title']} must match the pattern ${field['#pattern']}`
-    }
-    if (field['#type'] === 'email') {
-      validationStates.push('email')
-      validationMessages[
-        'email'
-      ] = `${field['#title']} must be a valid email address`
-    }
-    if (typeof field['#multiple'] === 'number') {
-      validationStates.push(`length:0,${field['#multiple']}`)
-      validationMessages['length'] =
-        field['#multiple_error'] ||
-        `More than ${field['#multiple']} selections are not allowed`
-    }
-
-    return {
-      validation: validationStates.join('|'),
-      validationMessages,
-      validationMeta
-    }
-  }
-
-  const getInputIcons = (
-    field
-  ): { prefixIcon?: string; suffixIcon?: string } => {
-    if (field['#field_prefix']) {
-      return {
-        prefixIcon: `icon-${field['#field_prefix']}`
-      }
-    }
-    return {}
-  }
-
+  const elements: ITideFormElement[] = webform?.elements || []
   const fields: any[] = []
 
   for (const [fieldKey, field] of Object.entries(elements)) {
@@ -101,7 +45,9 @@ const getFormSchemaFromMapping = async (
           help: field['#description'] || field['#help_title'],
           value: field['#default_value'],
           ...getValidation(field),
-          ...getInputIcons(field)
+          ...getInputIcons(field),
+          ...getMinMaxFields(field),
+          ...getCounterFields(field)
         }
         break
       case 'email':
@@ -115,7 +61,8 @@ const getFormSchemaFromMapping = async (
           help: field['#description'] || field['#help_title'],
           value: field['#default_value'],
           ...getValidation(field),
-          ...getInputIcons(field)
+          ...getInputIcons(field),
+          ...getMinMaxFields(field)
         }
         break
       case 'number':
@@ -143,7 +90,8 @@ const getFormSchemaFromMapping = async (
           help: field['#description'] || field['#help_title'],
           value: field['#default_value'],
           ...getValidation(field),
-          ...getInputIcons(field)
+          ...getInputIcons(field),
+          ...getMinMaxFields(field)
         }
         break
       case 'url':
@@ -157,7 +105,24 @@ const getFormSchemaFromMapping = async (
           help: field['#description'] || field['#help_title'],
           value: field['#default_value'],
           ...getValidation(field),
-          ...getInputIcons(field)
+          ...getInputIcons(field),
+          ...getMinMaxFields(field)
+        }
+        break
+      case 'textarea':
+        mappedField = {
+          $formkit: 'RplFormTextarea',
+          id: fieldKey,
+          name: fieldKey,
+          label: field['#title'],
+          disabled: field['#disabled'],
+          placeholder: field['#placeholder'],
+          rows: field['#rows'],
+          help: field['#description'] || field['#help_title'],
+          value: field['#default_value'],
+          ...getValidation(field),
+          ...getMinMaxFields(field),
+          ...getCounterFields(field)
         }
         break
       case 'checkbox':
@@ -236,11 +201,12 @@ const getFormSchemaFromMapping = async (
       }
       case 'webform_actions':
         mappedField = {
-          $formkit: 'RplFormSubmit',
+          $formkit: 'RplFormActions',
           name: 'submit',
           variant: 'filled',
           label: field['#submit__label'],
           id: fieldKey,
+          displayResetButton: !!webform?.settings?.form_reset,
           ...getInputIcons(field)
         }
         break
@@ -266,13 +232,27 @@ export const webformMapping = async (
   tidePageApi
 ): TideDynamicPageComponent<ITideWebform> => {
   return {
-    component: 'RplForm',
+    component: 'TideLandingPageWebForm',
     id: field.drupal_internal__id,
     title: field.field_paragraph_title,
     props: {
-      id: field?.field_paragraph_webform?.drupal_internal__id,
+      formId: field?.field_paragraph_webform?.drupal_internal__id,
+      hideFormOnSubmit:
+        field?.field_paragraph_webform?.settings?.confirmation_type ===
+        'inline',
+      successMessageTitle:
+        field?.field_paragraph_webform?.settings?.confirmation_title ||
+        'Form submitted',
+      successMessageHTML:
+        field?.field_paragraph_webform?.settings?.confirmation_message ||
+        'Thank you! Your response has been submitted.',
+      errorMessageTitle: 'Form not submitted',
+      errorMessageHTML:
+        field?.field_paragraph_webform?.settings
+          ?.submission_exception_message ||
+        'We are experiencing a server error. Please try again, otherwise contact us.',
       schema: await getFormSchemaFromMapping(
-        field?.field_paragraph_webform?.elements,
+        field?.field_paragraph_webform,
         tidePageApi
       )
     }