+++ packages/ripple-ui-forms/src/inputs/email.ts
@@ -3,6 +3,8 @@ import {
   createRplFormInput,
   defaultRplFormInputProps,
   inputLibrary,
+  minMaxProps,
+  minMaxRplFormProps,
   rplFeatures
 } from './input-utils'
 
@@ -18,6 +20,7 @@ export const email: FormKitTypeDefinition = {
     $cmp: 'RplFormInput',
     props: {
       ...defaultRplFormInputProps,
+      ...minMaxRplFormProps,
       type: 'email'
     }
   }),
@@ -34,7 +37,10 @@ export const email: FormKitTypeDefinition = {
   /**
    * An array of extra props to accept for this input.
    */
-  props: ['maxlength', 'minlength', 'placeholder', 'validationMeta'],
+  props: [...minMaxProps, 'placeholder', 'validationMeta'],
+  /**
+   * Forces node.props.type to be this explicit value.
+   */
   forceTypeProp: 'email',
   /**
    * Additional features that should be added to your input