+++ packages/ripple-ui-forms/src/inputs/url.ts
@@ -3,6 +3,8 @@ import {
   createRplFormInput,
   defaultRplFormInputProps,
   inputLibrary,
+  minMaxProps,
+  minMaxRplFormProps,
   rplFeatures
 } from './input-utils'
 
@@ -18,6 +20,7 @@ export const url: FormKitTypeDefinition = {
     $cmp: 'RplFormInput',
     props: {
       ...defaultRplFormInputProps,
+      ...minMaxRplFormProps,
       type: 'url'
     }
   }),
@@ -34,7 +37,7 @@ export const url: FormKitTypeDefinition = {
   /**
    * An array of extra props to accept for this input.
    */
-  props: ['placeholder', 'validationMeta'],
+  props: [...minMaxProps, 'placeholder', 'validationMeta'],
   /**
    * Forces node.props.type to be this explicit value.
    */